Updated git submodules
Project: genius master 6f8e7f2a9c0ff163c08427359d63630cc0dd8e67

Upgrading flag for full sync upgrades

This commit introduces a small utility that tracks whether or not
an upgrade is in progress. The purpose of this state is so that
code paths required during upgrade need not be entered during
regular application flow. The upgrade in progress state is toggled
via REST (or config xml). I tested it like this:

> curl -v -u admin:admin -X PUT -d @put-data -H \
"Content-Type: application/json" \
http://127.0.0.1:8181/restconf/config/genius-mdsalutil:config

...where the contents of the put-data file are:
{
"config": {
        "upgradeInProgress": false
         }
}

The REST call to retrieve the status is:

> curl -u admin:admin http://127.0.0.1:8181/restconf/config/genius-mdsalutil:config

Change-Id: I96d4255645b8650ee97154ff58c1afc52d1ac9f0
Signed-off-by: Josh <jhershbe@redhat.com>