Fix mis-advice for RESTCONF installation
There is no need for odl-netconf-connector-ssh
in normal RESTCONF scenarios.  This one case
is because odl-distribution-version is using
the config subsystem to expose information,
which in itself is a dated method.  In fact,
we have deprecated CSS this release.  At the
least, this updates the docs to recommend
a smaller feature set.  The other issues can
be addressed in later patches when they are
fixed in integration project

Change-Id: I154785d31f49f67a21fad31f87e9deac32cdab92
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>