Updated git submodules
Project: netvirt master cfac6b9839a14801629bfa1f07c6546ba69b16c7

Convert InterVpnLinkCache to non-static singleton

InterVpnLinkCache is now an interface with InterVpnLinkCacheImpl
that is advertised as a service. InterVpnLinkCache is also used
by fibmanager-impl which, unfortuantely, creates a circular
service dependency with vpnamnager-impl so the service import
ws made optional. A better sution would be to put the
InterVpnLinkCacheImpl in a separate bundle.

Change-Id: If09ba2b99110c1d43c87985110e69365ec76f33f
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>