Update patch set 2
Patch Set 2:

> Abandoned
 > 
 > all rework done on change 71288

Why create a new change? This erases valuable tracing of review history. How Gerrit determines if an existing patch should be updated or a new patch is updated is via the Change-Id field. DO NOT delete it when you are doing "git commit --amend" and Gerrit will do the right thing.

If for whatever reason you are re-writing the commit then ensure that the Change-Id field is copied over so that the correct commit is updated by Gerrit. Can we update this patch instead of the new one?

Patch-set: 2