Add Report Issue ribbon
Add a ribbon on the top right of the page which when clicked will
direct the reader to LF Jira to open an issue with the doc. It also
prefills the description with the version of dosc and current page
the user was viewing when they clicked the ribbon.

JIRA: DOCS-99
Change-Id: I74373ea862a123e69e3718eeb4329737ba2734e9
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>