Update patch set 1
Patch Set 1:

One thing I found with the docs project is when you click create it has a bunch of fields, particularly the "reporter" field that's required but it's confusing what to set it to. It should be automatic to be self but it's not when using this method.

I wonder if we should customize the docs project issue types and create a docs specific issue type before merging this patch to make the reporting a simple as possible.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0