Update patch set 1
Patch Set 1:

(1 comment)

> (1 comment)
 > 
 > One nit, otherwise looks good.

I'm just mechanically copying the dates forward. We can try to tweak them a bit if we need.

One idea I had was to edit the release start dates to always be Thursdays ("The Thursdays closet to 03/07 and 09/07)". As-is, they are fixed to a specific number day, so the day of the week will float. This time the release starts on a Friday, which causes all the other dates that are some whole number of weeks after the release start to also be on Fridays. Big events on Fridays are obviously not good for most folks, but especially our large group of devs in Israel.

Patch-set: 1
Label: Verified=0