Update patch set 1
Patch Set 1:

I was thinking we should always have two collumns in on this page. one for the dates of the current release (so keep Fluorine for now) and 
a column for the next release. Or even a third column to give a
little bit more idea of the future. But, at least with two releases listed it will be more clear that we are just on a simple 6m cadence.

no? yeah?

Patch-set: 1