Update patch set 2
Patch Set 2:

From our experience in Fluorine, do folks feel if we are micromanaging our Checkpoints?  Perhaps Checkpoints should not be a prescription of what actions to take, but rather that some set of deliveries have been finalized.  One example, would be to merge the MSI Integration, Version Bump, Version Bump Deadline, and CSIT into a single "Quarter Checkpoint".  Note that this isn't a date to which the actions should be performed (that's up to the individual projects to negotiate and advertise), only that all related actions should be completed before such date.

Patch-set: 2
Label: Verified=0