Update patch set 3
Patch Set 3:

Curious if others have an opinion here. This patch will "hide" the link to our docs. Instead of seeing the full link people will now just see "Documentation" and they won't know the full link until they click on it.

Personally I like to see the full like in this case in the README because we want folks to know where to look for our documentation if they want to rather than depend on the link from the README.

Patch-set: 3