Update patch set 9
Patch Set 9:

> From Fluorine, we do not have user/dev guides in landing page,
 > instead there are links to project specific documentation:
 > https://docs.opendaylight.org/en/stable-fluorine
 > To add your link in landing page follow this instruction sent some
 > time back: https://lists.opendaylight.org/pipermail/release/2018-July/015578.html

Ok to do it but still unclear to us, whether we shall or shall not do it : during the TSC call of the 30th of August, transportPCE was not mentionned when they evoked the "self-managed" project. And right now tpce is an unmanaged project. So prior to put the doc as a link we'd like to have a confirmation that transportPCE has been accepted for integration in Fluorine and that this link is also relevant for unmanaged projects.

Patch-set: 9
Reviewer: Gerrit User 7217 <7217@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0