Update patch set 17
Patch Set 17:

My git remote is renamed to gerrit.
If git remote is origin, then git review -d is not working.

Hence had to rename it to gerrit.

But the build fails due to git config --get remote.origin.url failure, as it only requests origin. Any solution for this?

Patch-set: 17
Label: Verified=0