Update patch set 1
Patch Set 1: Code-Review+2

While I'd prefer to migrate all the docs first before removing them as it gives us an incentive to actually do the work. Whereas deleting it just tells us to ignore it and not do it. It seems the community isn't interested in helping with such a migration so let's just delete everything in Neon. I'll spend my weekends migrating based on the fluorine docs.

Patch-set: 1
Label: Code-Review=+2