Update patch set 1
Patch Set 1:

(1 comment)

This is great, thanks for writing it up! I think there is one thing we need to clarify in this document: if each time supercommitter rights need to be activated under the cases listed the TSC is required to vote, or the TSC is only required to vote in the "Exceptional Cases"?

Patch-set: 1
Label: Code-Review=0