Update patch set 1
Patch Set 1:

> I don't think we should have merged this patch "Karaf" is NOT the
 > artifact we are telling end users to download. We need to wait for
 > the "opendaylight" artifact to be released!

Well I think merging this is better than leaving the links broken. Also I think we'll only get an opendaylight- distro if we do a MR+SM common distro, and it's not clear to me that will happen (I don't think we want this common distro long-term, and SM projects are broken short-term).

I think we should cherry-pick this to stable/fluorine, but I'll let you review.

Patch-set: 1