Update patch set 1
Patch Set 1:

Adding Luis, I think the current plan for now is to continue releasing the opendaylight artifact as the karaf artifact does not include the SM projects right?

Daniel suggested that we at least link to a working download in the meantime which sounds reasonable but I think we should have a clear path on which artifact we want to use moving forward.

Patch-set: 1