Update patch set 1
Patch Set 1:

> I do not think we need this right?

It's strange because the live HTML for latest and stable-neon shows the Neon link, but this rst on master does seem to need an update. Maybe we're missing a cherry-pick to master and the HTML is coming from stable/neon rst? or something else strange idk.

Patch-set: 1
Label: Code-Review=0