Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Not sure what to understand from that
> 
> Warning, treated as error:
> /w/workspace/docs-rtd-verify-magnesium/docs/release-notes/projects/transportpce.rst:document isn't included in any toctree
> ERROR: InvocationError for command /w/workspace/docs-rtd-verify-magnesium/.tox/docs/bin/sphinx-build -j auto -W -b html -n -d /w/workspace/docs-rtd-verify-magnesium/.tox/docs/tmp/doctrees ./docs/ /w/workspace/docs-rtd-verify-magnesium/docs/_build/html (exited with code 2)

docs for transportpce was moved to the transportpce project so we should be linking to transportpce. The current :ref: code is trying to link to non-existent docs in the docs project and not transportpce's docs.

I'll fix this shortly and update this patch with the new syntax.

Patch-set: 1
Label: Code-Review=0