Update patch set 1
Patch Set 1:

Hi Robert, I'm setting the docs link for MD-SAL to the 5.0.3 tag but was wondering is there plans to have a v5.0.x branch?

I see there's one for v4.0.x but none for the v5.0.x line and I think it would be better to set it to the branch so that docs updates are automatically picked up if any.

Patch-set: 1