Update patch set 1
Patch Set 1:

One issue I see is the v4.0.7 link has a version of 2.2.0 which means conf.yaml was not updated with the new version. I think hardcoding version info is bad practice and we've solved this issue in LF projects by having sphinx do a git tag lookup so the version is always equivalent to the tag. Or tag+devX to indicate a dev version which now many commits since last tag.

I'll find some time to port this logic into MD-SAL.

Patch-set: 1