Update patch set 4
Patch Set 4: -Code-Review

> Patch Set 4:
> 
> > (1 comment)
> 
> I personally liked more when it was 1 (https://docs.opendaylight.org/en/stable-neon/release-notes/index.html vs https://docs.opendaylight.org/en/stable-sodium/release-notes/index.html)

Luis you posted the same link. Here is what it looks like with maxdepth 1:

https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/docs-rtd-verify-sodium/122/html/release-notes/index.html

With depth 1 we only get a single link for "Sodium-SR1 Release Notes" but with depth 2 we get a list of projects as well. I'm fine with either, I do think depth 2 allows for more of an index than 1 which is only a single link.

Patch-set: 4