Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 2:
> > > > So I guess the answer to Jamo's question is no, we don't want these changes in Master?
> > > 
> > > We don't need them in master but would be worth cherry-picking anything relevant into stable/sodium and stable/fluorine maybe.
> > 
> > Got it.
> 
> FYI: It seems to me that it'd be easier if I just make the changes directly on each branch, instead of cherry-picking. Correct me if I am wrong, but once I do the CP, I need to take the previous changes and merge it into the new CP files.

It's up to you how you want to handle it. It really depends how in sync the branches are. Usually if branches are kept up to date then cherry-picking often results in trivial updates with no merge conflicts. If branches are not kept in sync.

(IE, regularly cherry-picking changes to each branch when a patch is merged, then merge conflicts might occur making it difficult)

In summary if you're having a lot of merge conflicts and think it will be easier to resolve by not cherry-picking then go that route.

Patch-set: 2