Update patch set 1
Patch Set 1: Code-Review-1

Not sure renaming all instances of YANG to Yang is correct. YANG is a acrynom and searching around the web it seems the standard is to write it in all caps as far as I can tell.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1