Remove docs branching scripts
Functionality of these scripts have been moved to releng/builder under
the autorelease project via I0aa6838d1fd4e2df536a4e30fb8751b7d3a466bf
to be handled as part of branch cutting activities for managed
projects.

There's the script docs_master_branch_update.sh needs to be rewritten
as it does not work for our purposes and in it's current state is quite
broken. We'll handle this as a separate task but removing here as this
script is not in use in production today.

Issue: DOCS-129
Signed-off-by: Thanh Ha <zxiiro@gmail.com>
Change-Id: If3f9b61636643c065ef11aa048debeb92422628e