Extract JIRA issues
We want to capture a table of JIRA issues for a number of MRI project
versions.

Nothing in upstream fits the bill, so we introduce our own Sphinx
extension, which uses the jira module to extract information from JIRA
and embed it into our RTD site.

Since there is no spell checker running on JIRA, we exclude files in
release-notes/projects/ from the Sphinx spell checker.

Change-Id: I5b0ac293dca2442776753078dbd1c79a52684677
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>