embed and fix metada of the technical charter pdf
from
https://www.opendaylight.org/wp-content/uploads/sites/14/2018/01/ODL-Technical-Charter-LF-Projects-LLC-FINAL.pdf

Title Metadata are displayed at the top of the browser Window.
They should reflect the content of the charter and not how it was produced.

Since contributors must comply with this technical charter
and its content is quoted by contributor guides,
it should preferably be embedded in the documentation to avoid dead link.

JIRA: TSC-296
Signed-off-by: Guillaume Lambert <guillaume.lambert@orange.com>
Change-Id: I21df665c285bda15a2981c4403ce2f3150fd1716